Updated typo

+++ covidprognosis/plmodules/xray_datamodule.py
@@ -123,7 +123,7 @@ class XrayDataModule(pl.LightningDataModule):
             transforms.Normalize(mean=mean, std=std)])
 
         self.train_dataset = torchvision.datasets.ImageFolder(os.path.join(dataset_dir+"train"),transform= transform_train)
-        self.train_dataset,self.val_dataset = torch.utils.data.random_split(self.train_dataset,[len(self.train_dataset)*0.80,len(self.train_dataset)*0.20])
+        self.train_dataset,self.val_dataset = torch.utils.data.random_split(self.train_dataset,[int(len(self.train_dataset)*0.80),int(len(self.train_dataset)*0.20]))
         self.test_dataset = torchvision.datasets.ImageFolder(os.path.join(dataset_dir+"tesst"),transform = transform_test)
